pyproject.toml
[tool.poetry]
name = "adk-data-science-chatbot"
version = "1.0.0"
description = "Google ADK Data Science Multi-Agent Chatbot System"
authors = ["Data Science Team"]
readme = "README.md"
packages = [{include = "app", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
websockets = "^12.0"
google-generativeai = "^0.3.2"
google-cloud-bigquery = "^3.13.0"
google-cloud-storage = "^2.10.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
pandas = "^2.1.4"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
scikit-learn = "^1.3.2"
plotly = "^5.17.0"
jupyter = "^1.0.0"
ipykernel = "^6.27.1"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
httpx = "^0.25.2"

[tool.poetry.group.test.dependencies]
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
pytest-timeout = "^2.2.0"
coverage = "^7.3.2"

[tool.poetry.scripts]
deploy = "deployment.deploy:main"
chat-server = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["app", "tests", "eval", "deployment"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.generativeai",
    "google.cloud.bigquery",
    "matplotlib.*",
    "seaborn.*",
    "plotly.*",
    "pandas.*",
    "numpy.*",
    "sklearn.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
    "eval"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "evaluation: marks tests as evaluation tests"
]
asyncio_mode = "auto"
python_files = [
    "test_*.py",
    "*_test.py"
]
python_classes = [
    "Test*"
]
python_functions = [
    "test_*"
]

[tool.coverage.run]
source = ["backend/app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "*/conftest.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]